Refactor: Remove a `cast` and tweak style (#22)
* Style: Run `nimpretty`

* Style: Use two spaces to indent

* Style: `os.createDir` -> `createDir`

In general, it's idiomatic to omit the module name in Nim.

* Style: Prefer camelCase for variable names

This is the usual naming convention. See:
- https://nim-lang.org/docs/nep1.html

* Style: Organise imports

It can be nice to separate stdlib imports from Nimble imports.

* Refactor: Tweak comments

* Refactor: Remove a `cast`

* Refactor: `c_signal` -> `posix.onSignal`

This does the same thing internally, but looks a bit simpler. It also
allows other signals to be added just by delimiting with a comma.